Bugfixes
git-svn-id: http://escidev5.fiz-karlsruhe.de:8888/prototype/eSciDocCore@38650 5c398c58-3d0a-0410-a38c-d0825f6816e0

+++ common/src/main/java/de/escidoc/core/common/business/fedora/resources/Container.java
@@ -316,12 +316,13 @@ public class Container extends GenericVersionableResourcePid
                 Datastream fedoraDs = null;
                 try {
                     fedoraDs = getMdRecord(nameInFedora);
-                }
-                catch (StreamNotFoundException e) {
+                    if(fedoraDs != null) {
+                        // FIXME remove the entire datastream
+                        fedoraDs.delete();
+                    }
+                } catch (StreamNotFoundException e) {
                     log.warn("Failed to set MdRecords.");
                 }
-                // FIXME remove the entire datastream
-                fedoraDs.delete();
             }
         }
